Drop Down Chapter Menu Change
Oddly, the two chapter drop down menus (at the top and bottom of a chapter) for stories had different appearances when you hovered over the choices. This normalizes them to the more readable of the two. This also keeps the selected (red) chapter the same color so you know where you are.